# docs/swagger.yaml
definitions:
  server.ABACPolicyDetails:
    type: object
...
  server.AttributeDefinitionDetails:
    type: object
...
  server.CreateABACPolicyRequest:
    type: object
  server.CreateABACPolicyResponse:
    type: object
  server.CreateAssignmentRequest:
    type: object
  server.CreateAssignmentResponse:
    type: object
  server.CreateAssociationRequest:
    type: object
  server.CreateAssociationResponse:
    type: object
  server.CreateAttributeDefinitionRequest:
    type: object
  server.CreateAttributeDefinitionResponse:
    type: object
  server.CreateObjectAttributeRequest:
    type: object
  server.CreateObjectAttributeResponse:
    type: object
...
  server.CreatePolicyClassRequest:
    type: object
  server.CreatePolicyClassResponse:
    type: object
...
  server.CreateUserAttributeRequest:
    type: object
  server.CreateUserAttributeResponse:
    type: object
  server.ErrorResponse:
    properties:
      error:
        description: |-
          Error code
          example: validation_error
        type: string
      message:
        description: |-
          Error message
          example: Role name is required
        type: string
    type: object
  server.NGACGraphResponse:
    type: object
...
  server.PolicyClassDetails:
    type: object
...
  server.UpdateABACPolicyRequest:
    type: object
  server.UpdateABACPolicyResponse:
    type: object
...
info:
  contact: {}
paths:
  /api/v1/abac/attributes:
    get:
      description: Get a list of all attribute definitions used in ABAC policies
      produces:
      - application/json
      responses:
        "200":
          description: List of attribute definitions
          schema:
            items:
              $ref: '#/definitions/server.AttributeDefinitionDetails'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List all attribute definitions
      tags:
      - 3-advanced-abac
    post:
      consumes:
      - application/json
      description: Create a new attribute definition for ABAC policies
      parameters:
      - description: Attribute definition creation request
        in: body
        name: attribute
        required: true
        schema:
          $ref: '#/definitions/server.CreateAttributeDefinitionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Attribute definition created successfully
          schema:
            $ref: '#/definitions/server.CreateAttributeDefinitionResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new attribute definition
      tags:
      - 3-advanced-abac
  /api/v1/abac/policies:
    get:
      description: Get a list of all Attribute-Based Access Control policies
      produces:
      - application/json
      responses:
        "200":
          description: List of ABAC policies
          schema:
            items:
              $ref: '#/definitions/server.ABACPolicyDetails'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List all ABAC policies
      tags:
      - 3-advanced-abac
    post:
      consumes:
      - application/json
      description: Create a new Attribute-Based Access Control policy
      parameters:
      - description: ABAC policy creation request
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/server.CreateABACPolicyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ABAC policy created successfully
          schema:
            $ref: '#/definitions/server.CreateABACPolicyResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new ABAC policy
      tags:
      - 3-advanced-abac
  /api/v1/abac/policies/{id}:
    delete:
      description: Delete an existing Attribute-Based Access Control policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ABAC policy deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Delete an ABAC policy
      tags:
      - 3-advanced-abac
    put:
      consumes:
      - application/json
      description: Update an existing Attribute-Based Access Control policy
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: string
      - description: ABAC policy update request
        in: body
        name: policy
        required: true
        schema:
          $ref: '#/definitions/server.UpdateABACPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ABAC policy updated successfully
          schema:
            $ref: '#/definitions/server.UpdateABACPolicyResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Policy not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Update an ABAC policy
      tags:
      - 3-advanced-abac
...
  /api/v1/ngac/assignments:
    post:
      consumes:
      - application/json
      description: Create a new assignment relationship in the NGAC system
      parameters:
      - description: Assignment creation request
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/server.CreateAssignmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Assignment created successfully
          schema:
            $ref: '#/definitions/server.CreateAssignmentResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new assignment
      tags:
      - 9-internal-ngac
  /api/v1/ngac/associations:
    post:
      consumes:
      - application/json
      description: Create a new association relationship in the NGAC system
      parameters:
      - description: Association creation request
        in: body
        name: association
        required: true
        schema:
          $ref: '#/definitions/server.CreateAssociationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Association created successfully
          schema:
            $ref: '#/definitions/server.CreateAssociationResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new association
      tags:
      - 9-internal-ngac
  /api/v1/ngac/graph:
    get:
      description: Get the complete NGAC graph showing all relationships and nodes
      produces:
      - application/json
      responses:
        "200":
          description: NGAC graph data
          schema:
            $ref: '#/definitions/server.NGACGraphResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Get the NGAC graph
      tags:
      - 9-internal-ngac
  /api/v1/ngac/object-attributes:
    post:
      consumes:
      - application/json
      description: Create a new object attribute in the NGAC system
      parameters:
      - description: Object attribute creation request
        in: body
        name: objectAttribute
        required: true
        schema:
          $ref: '#/definitions/server.CreateObjectAttributeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Object attribute created successfully
          schema:
            $ref: '#/definitions/server.CreateObjectAttributeResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new object attribute
      tags:
      - 9-internal-ngac
  /api/v1/ngac/policy-classes:
    get:
      description: Get a list of all policy classes in the NGAC system
      produces:
      - application/json
      responses:
        "200":
          description: List of policy classes
          schema:
            items:
              $ref: '#/definitions/server.PolicyClassDetails'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List all policy classes
      tags:
      - 9-internal-ngac
    post:
      consumes:
      - application/json
      description: Create a new policy class in the NGAC (Next Generation Access Control)
        system
      parameters:
      - description: Policy class creation request
        in: body
        name: policyClass
        required: true
        schema:
          $ref: '#/definitions/server.CreatePolicyClassRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Policy class created successfully
          schema:
            $ref: '#/definitions/server.CreatePolicyClassResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new policy class
      tags:
      - 9-internal-ngac
  /api/v1/ngac/user-attributes:
    post:
      consumes:
      - application/json
      description: Create a new user attribute in the NGAC system
      parameters:
      - description: User attribute creation request
        in: body
        name: userAttribute
        required: true
        schema:
          $ref: '#/definitions/server.CreateUserAttributeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User attribute created successfully
          schema:
            $ref: '#/definitions/server.CreateUserAttributeResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new user attribute
      tags:
      - 9-internal-ngac
...
swagger: "2.0"