# pkg/server/abac_handlers.go
package server

import (
	"encoding/json"
	"net/http"
)

// ABAC Handlers

// CreateABACPolicyHandler creates a new ABAC policy
// @Summary Create a new ABAC policy
// @Description Create a new Attribute-Based Access Control policy
// @Tags 3-advanced-abac
// @Accept json
// @Produce json
// @Param policy body CreateABACPolicyRequest true "ABAC policy creation request"
// @Success 201 {object} CreateABACPolicyResponse "ABAC policy created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/abac/policies [post]
func (h *BaseServer) CreateABACPolicyHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create ABAC policy not implemented"})
}

// GetABACPoliciesHandler retrieves all ABAC policies
// @Summary List all ABAC policies
// @Description Get a list of all Attribute-Based Access Control policies
// @Tags 3-advanced-abac
// @Produce json
// @Success 200 {array} ABACPolicyDetails "List of ABAC policies"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/abac/policies [get]
func (h *BaseServer) GetABACPoliciesHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Get ABAC policies not implemented"})
}

// UpdateABACPolicyHandler updates an existing ABAC policy
// @Summary Update an ABAC policy
// @Description Update an existing Attribute-Based Access Control policy
// @Tags 3-advanced-abac
// @Accept json
// @Produce json
// @Param id path string true "Policy ID"
// @Param policy body UpdateABACPolicyRequest true "ABAC policy update request"
// @Success 200 {object} UpdateABACPolicyResponse "ABAC policy updated successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 404 {object} ErrorResponse "Policy not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/abac/policies/{id} [put]
func (h *BaseServer) UpdateABACPolicyHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Update ABAC policy not implemented"})
}

// DeleteABACPolicyHandler deletes an ABAC policy
// @Summary Delete an ABAC policy
// @Description Delete an existing Attribute-Based Access Control policy
// @Tags 3-advanced-abac
// @Param id path string true "Policy ID"
// @Success 204 "ABAC policy deleted successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 404 {object} ErrorResponse "Policy not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/abac/policies/{id} [delete]
func (h *BaseServer) DeleteABACPolicyHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Delete ABAC policy not implemented"})
}

// CreateAttributeDefinitionHandler creates a new attribute definition
// @Summary Create a new attribute definition
// @Description Create a new attribute definition for ABAC policies
// @Tags 3-advanced-abac
// @Accept json
// @Produce json
// @Param attribute body CreateAttributeDefinitionRequest true "Attribute definition creation request"
// @Success 201 {object} CreateAttributeDefinitionResponse "Attribute definition created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/abac/attributes [post]
func (h *BaseServer) CreateAttributeDefinitionHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create attribute definition not implemented"})
}

// GetAttributeDefinitionsHandler retrieves all attribute definitions
// @Summary List all attribute definitions
// @Description Get a list of all attribute definitions used in ABAC policies
// @Tags 3-advanced-abac
// @Produce json
// @Success 200 {array} AttributeDefinitionDetails "List of attribute definitions"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/abac/attributes [get]
func (h *BaseServer) GetAttributeDefinitionsHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Get attribute definitions not implemented"})
}