# pkg/server/ngac_handlers.go
package server

import (
	"encoding/json"
	"net/http"
)

// NGAC Handlers

// CreatePolicyClassHandler creates a new policy class
// @Summary Create a new policy class
// @Description Create a new policy class in the NGAC (Next Generation Access Control) system
// @Tags 9-internal-ngac
// @Accept json
// @Produce json
// @Param policyClass body CreatePolicyClassRequest true "Policy class creation request"
// @Success 201 {object} CreatePolicyClassResponse "Policy class created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/ngac/policy-classes [post]
func (h *BaseServer) CreatePolicyClassHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create policy class not implemented"})
}

// CreateUserAttributeHandler creates a new user attribute
// @Summary Create a new user attribute
// @Description Create a new user attribute in the NGAC system
// @Tags 9-internal-ngac
// @Accept json
// @Produce json
// @Param userAttribute body CreateUserAttributeRequest true "User attribute creation request"
// @Success 201 {object} CreateUserAttributeResponse "User attribute created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/ngac/user-attributes [post]
func (h *BaseServer) CreateUserAttributeHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create user attribute not implemented"})
}

// CreateObjectAttributeHandler creates a new object attribute
// @Summary Create a new object attribute
// @Description Create a new object attribute in the NGAC system
// @Tags 9-internal-ngac
// @Accept json
// @Produce json
// @Param objectAttribute body CreateObjectAttributeRequest true "Object attribute creation request"
// @Success 201 {object} CreateObjectAttributeResponse "Object attribute created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/ngac/object-attributes [post]
func (h *BaseServer) CreateObjectAttributeHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create object attribute not implemented"})
}

// CreateAssignmentHandler creates a new assignment
// @Summary Create a new assignment
// @Description Create a new assignment relationship in the NGAC system
// @Tags 9-internal-ngac
// @Accept json
// @Produce json
// @Param assignment body CreateAssignmentRequest true "Assignment creation request"
// @Success 201 {object} CreateAssignmentResponse "Assignment created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/ngac/assignments [post]
func (h *BaseServer) CreateAssignmentHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create assignment not implemented"})
}

// CreateAssociationHandler creates a new association
// @Summary Create a new association
// @Description Create a new association relationship in the NGAC system
// @Tags 9-internal-ngac
// @Accept json
// @Produce json
// @Param association body CreateAssociationRequest true "Association creation request"
// @Success 201 {object} CreateAssociationResponse "Association created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/ngac/associations [post]
func (h *BaseServer) CreateAssociationHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create association not implemented"})
}

// GetPolicyClassesHandler retrieves all policy classes
// @Summary List all policy classes
// @Description Get a list of all policy classes in the NGAC system
// @Tags 9-internal-ngac
// @Produce json
// @Success 200 {array} PolicyClassDetails "List of policy classes"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/ngac/policy-classes [get]
func (h *BaseServer) GetPolicyClassesHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Get policy classes not implemented"})
}

// GetNGACGraphHandler retrieves the NGAC graph
// @Summary Get the NGAC graph
// @Description Get the complete NGAC graph showing all relationships and nodes
// @Tags 9-internal-ngac
// @Produce json
// @Success 200 {object} NGACGraphResponse "NGAC graph data"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/ngac/graph [get]
func (h *BaseServer) GetNGACGraphHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Get NGAC graph not implemented"})
}